+++ src/cmd/mod.rs
@@ -1,5 +1,6 @@
 pub mod auth;
 pub mod bastion;
+pub mod environment;
 pub mod fleet;
 pub mod host;
 pub mod lifecycle;